+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Master
+
+- SERVER: Updated .TC (GH-228) TLD definition.
+
+- NEW: Added full whois.meridiantld.net parser (GH-228).
+
 
 ## Release 3.1.2