+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - NEW: Added full whois.meridiantld.net parser (GH-228).
 
+- NEW: Added full whois.nic.gd parser (GH-227).
+
 
 ## Release 3.1.2